pyproject.toml
[tool.poetry]
name = "tests"
version = "0.1.0"
description = ""
authors = ["dharun-karthick-ramaraj <dharun.ramaraj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
selenium = "^4.26.1"
boto3 = "^1.35.68"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
certifi = "^2024.8.30"
snowflake-connector-python = "^3.12.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
pinecone-client = "^5.0.1"
langchain-pinecone = "^0.2.0"
langchain-nvidia-ai-endpoints = "^0.3.5"
langchain = "^0.3.9"
langchain-openai = "^0.2.10"
langgraph = "^0.2.53"
langchain-core = "^0.3.21"
pytest = "^8.3.4"
apache-airflow = "^2.10.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
